feat: move analytics code out of delayed
